# openapi.yaml
openapi: 3.0.0
info:
  title: HeyGen MCP Adapter API
  version: v1.0.0
  description: API for the HeyGen MCP Adapter, facilitating video generation and status checks with HeyGen.
servers:
  - url: http://localhost:3001 # Example URL, will be replaced by actual deployment URL
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /video/generate:
    post:
      summary: Generate a new video
      description: Submits a script to HeyGen for video generation using a specified avatar and voice.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                script_text:
                  type: string
                  description: The script content for the video.
                  example: "Hello world, this is a test video."
                  minLength: 1
                  maxLength: 5000
                avatar_id:
                  type: string
                  description: The ID of the HeyGen avatar to use.
                  example: "avatar_default_id"
                  minLength: 1
                  maxLength: 100
                  pattern: '^[A-Za-z0-9_-]+$'
                voice_id:
                  type: string
                  description: The ID of the HeyGen voice to use.
                  example: "voice_default_id"
                  minLength: 1
                  maxLength: 100
                  pattern: '^[A-Za-z0-9_-]+$'
              required:
                - script_text
                - avatar_id
                - voice_id
      responses:
        '202':
          description: Video generation request accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: string
                    description: The unique identifier for the HeyGen video being generated.
                    example: "heygen_video_abc123"
                  status:
                    type: string
                    description: Initial status of the video generation process.
                    example: "PENDING"
        '400':
          description: Bad Request - Invalid input parameters.
        '500':
          description: Internal Server Error - Failed to initiate video generation.

  /video/{video_id}/status:
    get:
      summary: Get video status
      description: Retrieves the current status of a video generation task and the video URL if available.
      security:
        - bearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          description: The unique identifier of the HeyGen video.
          schema:
            type: string
            example: "heygen_video_abc123"
      responses:
        '200':
          description: Successfully retrieved video status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: string
                    description: The unique identifier for the HeyGen video.
                  status:
                    type: string
                    description: Current status of the video (e.g., PENDING, PROCESSING, READY, FAILED).
                    example: "READY"
                  video_url:
                    type: string
                    format: url
                    description: URL to the generated video MP4 file (present if status is READY).
                    nullable: true
                    example: "https://heygen-videos.com/video_abc123.mp4"
                  error_message:
                    type: string
                    description: Error message if the video generation failed.
                    nullable: true
                required:
                  - video_id
                  - status
        '404':
          description: Not Found - Video ID not found.
        '500':
          description: Internal Server Error.